requirements.txt
Django==5.0.2
python-dotenv==1.0.0
django-crispy-forms==2.1
crispy-bootstrap4==2023.1
django-recaptcha==3.0.0
mysqlclient==2.2.4
gunicorn==21.2.0
